Move headers in a subfolder to help integration in another project. (#12)
Add the possibility to send a NOP command (useful for debug purpose or to workaround bad NIC behavior).
Fix: if link layer believed  that a frame was missing and was wrong, datagraéms can be associated to wrong handler.

+++ src/CoE.cc
@@ -42,7 +42,7 @@ namespace kickcat
         uint8_t* pos = reinterpret_cast<uint8_t*>(data);
         size = *data_size;
         uint32_t already_read = 0;
-        for (int32_t i = 1; i <= object_size; ++i)
+        for (uint8_t i = 1; i <= object_size; ++i)
         {
             size = *data_size - already_read;
             if (size == 0)
